Search to Product description intent call completed by Mahendran

+++ Mcom/app/src/main/java/com/mindmade/mcom/utilclasses/model/SearchModel.java
@@ -28,7 +28,7 @@ public class SearchModel {
         @SerializedName(Const.TITLE_KEY)
         String name;
         @SerializedName(Const.ID_KEY)
-        float id;
+        long id;
 
         public String getName() {
             return name;
@@ -38,11 +38,11 @@ public class SearchModel {
             this.name = name;
         }
 
-        public float getId() {
+        public long getId() {
             return id;
         }
 
-        public void setId(float id) {
+        public void setId(long id) {
             this.id = id;
         }
     }
